parse html in bootstrap period
